backend/requirements.txt
flask
flask-cors
shodan
python-dotenv
flask-sqlalchemy
flask-jwt-extended
bcrypt
python-dotenv
psycopg2-binary